(PC-8100): Mutualize spectree security scheme declaration tool
