+++ src/pcapi/routes/pro/blueprints.py
@@ -1,8 +1,7 @@
 from flask import Blueprint
-from spectree import SpecTree
 
+from pcapi.serialization.spec_tree import ExtendedSpecTree
 from pcapi.serialization.utils import before_handler
-from pcapi.validation.routes import users_authentifications
 
 
 pro_api_v2 = Blueprint("pro_api_v2", __name__)
@@ -10,35 +9,13 @@ pro_api_v2 = Blueprint("pro_api_v2", __name__)
 
 API_KEY_AUTH = "ApiKeyAuth"
 
-
-class ProApiV1BluePrints(SpecTree):
-    def _generate_spec(self) -> dict:
-        schema = super()._generate_spec()
-        schema["components"]["securitySchemes"] = {
-            API_KEY_AUTH: {
-                "type": "http",
-                "scheme": "bearer",
-            }
-        }
-
-        # Find routes that are protected by authentication
-        # We may drop this whenever Spectree supports authentication
-        for route in self.backend.find_routes():
-            path, _parameters = self.backend.parse_path(route)
-            if path not in schema["paths"]:
-                continue
-            for method, func in self.backend.parse_func(route):
-                if method.lower() not in schema["paths"][path]:
-                    continue
-                if self.backend.bypass(func, method) or self.bypass(func):
-                    continue
-
-                if getattr(func, "requires_authentication", None):
-                    if users_authentifications.API_KEY in func.requires_authentication:
-                        schema["paths"][path][method.lower()]["security"] = [{API_KEY_AUTH: []}]
-
-        return schema
+security_schemes = {
+    API_KEY_AUTH: {
+        "type": "http",
+        "scheme": "bearer",
+    }
+}
 
 
-api = ProApiV1BluePrints("flask", MODE="strict", before=before_handler, PATH="/")
+api = ExtendedSpecTree("flask", MODE="strict", before=before_handler, PATH="/", security_schemes=security_schemes)
 api.register(pro_api_v2)
